third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/script-charset-02-expected.txt

This is a testharness.js-based test.
PASS Script @type: unknown parameters 
FAIL Script @type: unknown parameters 1 assert_equals: expected 5 but got 10
Harness: the test ran to completion.
